// src/lib/types.ts
export interface SelectedCustomization {
  type: string;
  label: string;
  selectedValue: string;
  priceModifier: number;
}

export interface OrderItem {
  product: string;
  productName: string;
  productImage: string;
  customizations: SelectedCustomization[];
  quantity: number;
  unitPrice: number;
}

export interface ShippingAddress {
  fullName: string;
  phone: string;
  street: string;
  city: string;
  province: string;
  zip: string;
}

export interface Order {
  _id: string;
  orderNumber: string;
  user: { _id: string; name: string; email: string } | string;
  items: OrderItem[];
  orderType: "regular" | "pre-order" | "custom";
  shippingAddress: ShippingAddress;
  paymentMethod: "stripe" | "paypal" | "gcash" | "cod";
  paymentStatus: "unpaid" | "deposit_paid" | "fully_paid" | "cod_pending" | "paid";
  orderStatus: "pending" | "confirmed" | "processing" | "shipped" | "completed" | "cancelled";
  totalAmount: number;
  depositAmount: number;
  balanceDue: number;
  notes: string;
  referenceImage?: string;
  designDescription?: string;
  customApprovalStatus?: "none" | "pending" | "approved" | "rejected";
  customRejectionReason?: string;
  courierName?: string;
  trackingNumber?: string;
  assignedTo: { _id: string; name: string; email?: string } | string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PaginationInfo {
  page: number;
  limit: number;
  total: number;
  pages: number;
}

export interface User {
  _id: string;
  name: string;
  email: string;
  role: "admin" | "staff" | "customer";
  phone: string;
  address: {
    street: string;
    city: string;
    province: string;
    zip: string;
  };
  isActive: boolean;
  createdAt: string;
}